Add support for processing only specific indexes

+++ resolwe/elastic/management/commands/elastic_purge.py
@@ -10,13 +10,23 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 from django.core.management.base import BaseCommand
 
 from resolwe.elastic.builder import index_builder
+from resolwe.elastic.mixins import ElasticIndexFilterMixin
 
 
-class Command(BaseCommand):
+class Command(ElasticIndexFilterMixin, BaseCommand):
     """Purge ElasticSearch indexes."""
 
     help = "Purge ElasticSearch indexes."
 
+    def handle_index(self, index):
+        """Process index."""
+        index.destroy()
+        index.create_mapping()
+
     def handle(self, *args, **options):
         """Command handle."""
-        index_builder.delete()
+        if self.has_filter(options):
+            self.filter_indices(options)
+        else:
+            # Process all indices.
+            index_builder.delete()
